feat: added reviews functionality

+++ src/main/java/ru/yandex/practicum/filmorate/dao/ReviewsRepository.java
@@ -6,14 +6,12 @@ import java.util.Collection;
 
 public interface ReviewsRepository {
     Review addReview(Review review);
-
-    Collection<Review> getAllReviews(int filmId, int count);
-
+    Collection<Review> getAllReviews(long filmId, int count);
     Review updateReview(Review review);
-    void removeReviewById(int id);
-    Review getReviewById(int id);
-    void addLike(int reviewId, long userId);
-    void addDislike(int reviewId, long userId);
-    void removeLike(int reviewId, long userId);
-    void removeDislike(int reviewId, long userId);
+    void removeReviewById(long id);
+    Review getReviewById(long id);
+    void addLike(long reviewId, long userId);
+    void addDislike(long reviewId, long userId);
+    void removeLike(long reviewId, long userId);
+    void removeDislike(long reviewId, long userId);
 }
